#76 [feat] : chat api 서버 통신 모델, 뷰모델

+++ app/src/main/java/com/sopt/withsuhyeon/domain/entity/ChatRoomInfoModel.kt
@@ -4,9 +4,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ChatRoomInfoModel(
-    val postId: Int,
-    val ownerId: Int,
-    val writerId: Int,
+    val postId: Long,
+    val ownerId: Long,
+    val writerId: Long,
     val ownerChatRoomId: String,
     val peerChatRoomId: String
 )
